requirements.txt
torch              1.9.1+cu111
torchaudio         0.9.1
torchvision        0.10.1+cu111
tqdm               4.64.1
typing_extensions  4.7.1
matplotlib         3.5.3
Pillow             9.3.0
pandas             1.3.5
scipy              1.2.0
numpy              1.21.6
opencv-python      4.8.1.78
scikit-image       0.19.2
scikit-learn       1.0.2